+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.2] - 16/04/2026
+
+* [refactor] Move utility functions to src/helpers.py
+* [feature] Add check_event_chain() helper for event inspection via action_id
+
 ## [0.0.1] - 16/04/2026
 
 * First commit: baseline of Transition Control project